+++ CHANGELOG.md
@@ -26,6 +26,10 @@
   - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#072)
   - [diff](https://github.com/getsentry/sentry-native/compare/0.7.0...0.7.2)
 
+### Dependencies
+
+- Bump Spring Boot from 2.7.5 to 2.7.18
+
 ## 7.10.0
 
 ### Features