+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Android: Report thermal device events as breadcrumbs ([#3547](https://github.com/getsentry/sentry-java/pull/3547))
+
 ## 7.11.0
 
 ### Features
 
 - Report dropped spans ([#3528](https://github.com/getsentry/sentry-java/pull/3528))
-- Android: Report thermal device events as breadcrumbs ([#3547](https://github.com/getsentry/sentry-java/pull/3547))
 
 ### Fixes